stats: implementation of metrics calculations has been moved to a separate metrics package
